requirements.txt
streamlit==1.35.0
tensorflow-cpu==2.12.0
numpy==1.23.5
pillow==9.5.0
h5py==3.8.0